+++ text-ui-test/runtest.sh
@@ -10,12 +10,12 @@ fi
 if [ -e "./ACTUAL.TXT" ]
 then
     rm ACTUAL.TXT
+    rm duke.txt
 fi
 
-rm duke.txt
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/duke/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
